don't buffer reading during connection bootstrap
(this was identified due to flakey Snappy tests)

if nsqd sends the IDENTIFY + handshake packets together
then go-nsq reads and buffers both, preventing an UpgradeX()
from being able to access the handshake data because we were
wrapping the underlying connection and not the buffered one.

fix this by not buffering at all during connection setup
and enabling it before handoff.