consumer: fix panic in nextLookupdEndpoint
If you use " myhostname:4161" as NSQd HTTP address
then the code panics in nextLookupdEndpoint:

    panic: parse "http:// myhostname:4161": invalid character " " in host name

This is caused different validation of the input when configured
in ConnectToNSQLookupd and when used in nextLookupdEndpoint.

Moving preparation of the URL to the beginning so that we don't
need to do any input validation in nextLookupdEndpoint.

Fixes https://github.com/nsqio/go-nsq/issues/316